Move collecting related methods to a separate module
