+++ crates/forge/src/test_crate_summary.rs
@@ -1,5 +1,6 @@
+use crate::collecting::TestCrateType;
 use crate::test_case_summary::TestCaseSummary;
-use crate::{RunnerStatus, TestCrateType};
+use crate::RunnerStatus;
 
 /// Summary of the test run in the file
 #[derive(Debug, PartialEq, Clone)]
